reflactor: remove appstate to common crate
