+++ mc-common/src/lib.rs
@@ -1,3 +1,4 @@
 extern crate core;
 
 pub mod types;
+pub mod app;
