--- mc-registry/src/cache/mod.rs
@@ -1,1 +1,0 @@
-pub mod mcp_servers;
