+++ mc-registry/src/mcp_server.rs
@@ -1,11 +1,11 @@
-use crate::event::Event;
-use crate::{AppState, Response};
 use axum::Json;
 use axum::extract::{Query, State};
 use axum::http::StatusCode;
 use mc_db::model::{CreateFrom, McpServers, SettingKey};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
+use mc_common::app::{AppState, Response};
+use mc_common::app::event::Event;
 
 #[derive(Deserialize, Debug)]
 pub struct ListAllRequest {
