+++ mc-service/src/reverse_proxy/message.rs
@@ -9,7 +9,6 @@ use hyper::body::Frame;
 use hyper_rustls::HttpsConnector;
 use hyper_util::client::legacy::Client;
 use hyper_util::client::legacy::connect::HttpConnector;
-use mc_registry::cache::mcp_servers::{Cache, McpServerInfo};
 use once_cell::sync::Lazy;
 use regex::Regex;
 use std::convert::Infallible;
@@ -19,6 +18,7 @@ use std::task::Poll;
 use tokio_stream::StreamExt;
 use tokio_stream::wrappers::ReceiverStream;
 use tower_service::Service;
+use mc_common::app::cache::{Cache, McpServerInfo};
 
 static REGEX_MESSAGE_ROUTER: Lazy<Regex> =
     Lazy::new(|| Regex::new(r"^/proxy/message/([^/]+)/([^/]+)(/.*)?$").unwrap());
