+++ mc-service/src/server.rs
@@ -1,16 +1,20 @@
 use crate::config::{AppConfig, McpRegistry};
 use crate::reverse_proxy::connection::ConnectionService;
 use crate::reverse_proxy::message::MessageService;
-use axum::Router;
+use axum::extract::{Request, State};
+use axum::middleware::Next;
+use axum::response::Response;
 use axum::routing::{get, post};
+use axum::{Router, middleware};
+use http::StatusCode;
 use hyper_rustls::HttpsConnectorBuilder;
 use hyper_util::client::legacy::Client;
 use hyper_util::rt::TokioExecutor;
 use mc_booter::app::application::Application;
+use mc_common::app::cache::Cache;
+use mc_common::app::event::Event;
+use mc_common::app::{AppState, HandlerManager};
 use mc_db::DBClient;
-use mc_registry::cache::mcp_servers::Cache;
-use mc_registry::event::Event;
-use mc_registry::{AppState, HandlerManager};
 use std::error::Error;
 use std::sync::Arc;
 use tokio::runtime::Runtime;
@@ -79,6 +83,10 @@ impl McpCenterServer {
                 "/api/registry/mcp-server",
                 post(mc_registry::register_mcp_server),
             )
+            .layer(middleware::from_fn_with_state(
+                (self.config.clone(), state.clone()),
+                check_api_key,
+            ))
             .with_state(state);
 
         // starting axum service
@@ -163,7 +171,8 @@ impl Application for McpCenterServer {
 
         let manager = HandlerManager::new(db_client.clone())
             .with_mcp_handler()
-            .with_system_settings_handler();
+            .with_system_settings_handler()
+            .with_api_keys_handler();
 
         let state = AppState::new(
             db_client.clone(),
@@ -193,4 +202,17 @@ fn build_external_api_registry(url: String, token: Option<String>) -> Registry {
         authorization: token,
     };
     Registry::ExternalAPI(config)
+}
+
+async fn check_api_key(
+    State((config, state)): State<(AppConfig, AppState)>,
+    req: Request,
+    next: Next,
+) -> Result<Response, StatusCode> {
+    if let Some(key) = req.headers().get(http::header::AUTHORIZATION) {
+        if key == config.mcp_center.admin_token.as_str() {
+            return Ok(next.run(req).await);
+        }
+    }
+    Err(StatusCode::UNAUTHORIZED)
 }
